checking roles to show respective ui elements

+++ resources/views/admin/gad-plan-list/index.blade.php
@@ -14,40 +14,58 @@
                 <div class="card">
                     <div class="card-header">
                         <i class="fa fa-align-justify"></i> {{ trans('admin.gad-plan-list.actions.index') }}
-                        <a class="btn btn-primary btn-spinner btn-sm pull-right m-b-0" href="{{ url('admin/gad-plan-lists/create') }}" role="button"><i class="fa fa-plus"></i>&nbsp; {{ trans('admin.gad-plan-list.actions.create') }}</a>
-                    </div>
+
+                        
+                        @if(Auth::user()->roles()->pluck('id')[0] == 2)
+                            <a class="btn btn-primary btn-spinner btn-sm pull-right m-b-0" href="{{ url('admin/gad-plan-lists/create') }}" role="button"><i class="fa fa-plus"></i>&nbsp; {{ trans('admin.gad-plan-list.actions.create') }}</a>
+                        @else
+                            <form @submit.prevent="deleteItem(item.resource_url)">
+                                <button type="submit" class="btn btn-success btn-spinner btn-sm pull-right m-b-0 text-white ml-2" title="{{ trans('brackets/admin-ui::admin.btn.accept') }}" role="button"><i class="fa fa-check"></i>&nbsp; {{ trans('admin.gad-plan-list.actions.accept') }}</button>
+                            </form>
+                            <form @submit.prevent="deleteItem(item.resource_url)">
+                                <button type="submit" class="btn btn-danger btn-spinner btn-sm pull-right m-b-0 text-white" title="{{ trans('brackets/admin-ui::admin.btn.accept') }}" role="button"><i class="fa fa-close"></i>&nbsp; {{ trans('admin.gad-plan-list.actions.decline') }}</button>
+                            </form>
+                        @endif
+                        </div>
                     <div class="card-body" v-cloak>
                         <div class="card-block">
-                            <form @submit.prevent="">
-                                <div class="row justify-content-md-between">
-                                    <div class="col col-lg-7 col-xl-5 form-group">
-                                        <div class="input-group">
-                                            <input class="form-control" placeholder="{{ trans('brackets/admin-ui::admin.placeholder.search') }}" v-model="search" @keyup.enter="filter('search', $event.target.value)" />
-                                            <span class="input-group-append">
-                                                <button type="button" class="btn btn-primary" @click="filter('search', search)"><i class="fa fa-search"></i>&nbsp; {{ trans('brackets/admin-ui::admin.btn.search') }}</button>
-                                            </span>
+
+                            
+                            @if(Auth::user()->roles()->pluck('id')[0] == 2)
+                                <form @submit.prevent="">
+                                    <div class="row justify-content-md-between">
+                                        <div class="col col-lg-7 col-xl-5 form-group">
+                                            <div class="input-group">
+                                                <input class="form-control" placeholder="{{ trans('brackets/admin-ui::admin.placeholder.search') }}" v-model="search" @keyup.enter="filter('search', $event.target.value)" />
+                                                <span class="input-group-append">
+                                                    <button type="button" class="btn btn-primary" @click="filter('search', search)"><i class="fa fa-search"></i>&nbsp; {{ trans('brackets/admin-ui::admin.btn.search') }}</button>
+                                                </span>
+                                            </div>
+                                        </div>
+                                        <div class="col-sm-auto form-group ">
+                                            <select class="form-control" v-model="pagination.state.per_page">
+                                                
+                                                <option value="10">10</option>
+                                                <option value="25">25</option>
+                                                <option value="100">100</option>
+                                            </select>
                                         </div>
                                     </div>
-                                    <div class="col-sm-auto form-group ">
-                                        <select class="form-control" v-model="pagination.state.per_page">
-                                            
-                                            <option value="10">10</option>
-                                            <option value="25">25</option>
-                                            <option value="100">100</option>
-                                        </select>
-                                    </div>
-                                </div>
-                            </form>
-
+                                </form>
+                            @endif
                             <table class="table table-hover table-listing">
                                 <thead>
                                     <tr>
-                                        <th class="bulk-checkbox">
-                                            <input class="form-check-input" id="enabled" type="checkbox" v-model="isClickedAll" v-validate="''" data-vv-name="enabled"  name="enabled_fake_element" @click="onBulkItemsClickedAllWithPagination()">
-                                            <label class="form-check-label" for="enabled">
-                                                #
-                                            </label>
-                                        </th>
+
+                                        
+                                        @if(Auth::user()->roles()->pluck('id')[0] == 2)
+                                            <th class="bulk-checkbox">
+                                                <input class="form-check-input" id="enabled" type="checkbox" v-model="isClickedAll" v-validate="''" data-vv-name="enabled"  name="enabled_fake_element" @click="onBulkItemsClickedAllWithPagination()">
+                                                <label class="form-check-label" for="enabled">
+                                                    #
+                                                </label>
+                                            </th>
+                                        @endif
 
                                         <th is='sortable' :column="'id'">{{ trans('admin.gad-plan-list.columns.id') }}</th>
                                         <th is='sortable' :column="'gad_issue_mandate'">{{ trans('admin.gad-plan-list.columns.gad_issue_mandate') }}</th>
@@ -76,11 +94,14 @@
                                 </thead>
                                 <tbody>
                                     <tr v-for="(item, index) in collection" :key="item.id" :class="bulkItems[item.id] ? 'bg-bulk' : ''">
-                                        <td class="bulk-checkbox">
-                                            <input class="form-check-input" :id="'enabled' + item.id" type="checkbox" v-model="bulkItems[item.id]" v-validate="''" :data-vv-name="'enabled' + item.id"  :name="'enabled' + item.id + '_fake_element'" @click="onBulkItemClicked(item.id)" :disabled="bulkCheckingAllLoader">
-                                            <label class="form-check-label" :for="'enabled' + item.id">
-                                            </label>
-                                        </td>
+                                        
+                                        @if(Auth::user()->roles()->pluck('id')[0] == 2)
+                                            <td class="bulk-checkbox">
+                                                <input class="form-check-input" :id="'enabled' + item.id" type="checkbox" v-model="bulkItems[item.id]" v-validate="''" :data-vv-name="'enabled' + item.id"  :name="'enabled' + item.id + '_fake_element'" @click="onBulkItemClicked(item.id)" :disabled="bulkCheckingAllLoader">
+                                                <label class="form-check-label" :for="'enabled' + item.id">
+                                                </label>
+                                            </td>
+                                        @endif
 
                                         <td>@{{ item.id }}</td>
                                         <td>@{{ item.gad_issue_mandate }}</td>
@@ -93,20 +114,25 @@
                                         <td>@{{ item.sourceofbudget.name }}</td>
                                         <td>@{{ item.responsible_unit }}</td>
                                         
-                                        <td>
-                                            <div class="row no-gutters">
-                                                <div class="col-auto">
-                                                    <a class="btn btn-sm btn-spinner btn-info" :href="item.resource_url + '/edit'" title="{{ trans('brackets/admin-ui::admin.btn.edit') }}" role="button"><i class="fa fa-edit"></i></a>
+
+                                        @if(Auth::user()->roles()->pluck('id')[0] == 2)
+                                            <td>
+                                                <div class="row no-gutters">
+                                                    <div class="col-auto">
+                                                        <a class="btn btn-sm btn-spinner btn-info" :href="item.resource_url + '/edit'" title="{{ trans('brackets/admin-ui::admin.btn.edit') }}" role="button"><i class="fa fa-edit"></i></a>
+                                                    </div>
+                                                    <form class="col" @submit.prevent="deleteItem(item.resource_url)">
+                                                        <button type="submit" class="btn btn-sm btn-danger" title="{{ trans('brackets/admin-ui::admin.btn.delete') }}"><i class="fa fa-trash-o"></i></button>
+                                                    </form>
                                                 </div>
-                                                <form class="col" @submit.prevent="deleteItem(item.resource_url)">
-                                                    <button type="submit" class="btn btn-sm btn-danger" title="{{ trans('brackets/admin-ui::admin.btn.delete') }}"><i class="fa fa-trash-o"></i></button>
-                                                </form>
-                                            </div>
-                                        </td>
+                                            </td>
+
+                                        @endif
                                     </tr>
                                 </tbody>
                             </table>
-
+                            
+                            @if(Auth::user()->roles()->pluck('id')[0] == 2)
                             <div class="row" v-if="pagination.state.total > 0">
                                 <div class="col-sm">
                                     <span class="pagination-caption">{{ trans('brackets/admin-ui::admin.pagination.overview') }}</span>
@@ -115,7 +141,7 @@
                                     <pagination></pagination>
                                 </div>
                             </div>
-
+                            @endif
                             <div class="no-items-found" v-if="!collection.length > 0">
                                 <i class="icon-magnifier"></i>
                                 <h3>{{ trans('brackets/admin-ui::admin.index.no_items') }}</h3>
